refactor(settings/copilot): 将 Copilot 设置重构为目录模块结构
