+++ src/renderer/pages/settings/copilot/index.tsx
@@ -1,11 +1,5 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React from 'react';
-import CopilotModalContent from '@/renderer/components/SettingsModal/contents/CopilotModalContent';
+import CopilotModalContent from '@/renderer/pages/settings/copilot/components/CopilotModalContent';
 import PageWrapper from '@renderer/components/base/PageWrapper';
 
 const CopilotSettings: React.FC = () => {
